+++ duckscript_sdk/src/sdk/std/on_error/trigger_error/mod.rs
@@ -1,0 +1,45 @@
+use crate::utils::pckg;
+use duckscript::types::command::{Command, CommandResult};
+
+#[cfg(test)]
+#[path = "./mod_test.rs"]
+mod mod_test;
+
+#[derive(Clone)]
+pub(crate) struct CommandImpl {
+    package: String,
+}
+
+impl Command for CommandImpl {
+    fn name(&self) -> String {
+        pckg::concat(&self.package, "TriggerError")
+    }
+
+    fn aliases(&self) -> Vec<String> {
+        vec!["trigger_error".to_string()]
+    }
+
+    fn help(&self) -> String {
+        include_str!("help.md").to_string()
+    }
+
+    fn clone_and_box(&self) -> Box<dyn Command> {
+        Box::new((*self).clone())
+    }
+
+    fn run(&self, arguments: Vec<String>) -> CommandResult {
+        let message = if arguments.is_empty() {
+            "Error".to_string()
+        } else {
+            arguments[0].clone()
+        };
+
+        CommandResult::Error(message)
+    }
+}
+
+pub(crate) fn create(package: &str) -> Box<dyn Command> {
+    Box::new(CommandImpl {
+        package: package.to_string(),
+    })
+}